DPC-80: Now testing all queue types
We’re now running the integration tests for both the memory and distributed queue types.

+++ dpc-aggregation/src/main/java/gov/cms/dpc/aggregation/qclient/AbstractMockQueueClient.java
@@ -25,7 +25,7 @@ public abstract class AbstractMockQueueClient implements JobQueue {
         // TODO
     }
 
-    public int queueSize() {
+    public long queueSize() {
         // TODO
         return -1;
     }
